Move the dummy storage service to the `decidim-dev` module
